+++ codegrade/docs/about/changelog.rst
@@ -3,6 +3,9 @@ Changelog
 
 .. deprecation_note:: /changelog
 
+.. metadata::
+   priority = 0.0
+   ignore = false
 
 Version *Next*
 --------------
@@ -27,6 +30,9 @@ Version Nobel.2
 
 .. json_include:: ../../release_notes.json
    :path: Nobel.2/message
+
+.. raw::
+   <iframe width="560" height="315" src="https://www.youtube.com/embed/qZGwRz8wnSY" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 
 Features
 ^^^^^^^^^